Avoid compiling reqd_sub_group_size kernels larger than max
In some cases, the device compilers may not be able to adhere to a
requested sub-group size, despite the device supporting it. This could
for example be when instructions are used that cannot be done for
certain sub-group sizes. To check this, the
`info::kernel_device_specific::max_sub_group_size` info query can be
requested. However, since the query can only be done on kernels, the
kernel needs to be compiled first. As such, this commit changes cases
that use `reqd_sub_group_size` to first check the max sub-group size of
as replicated kernel without the attribute and avoid building the
kernel with the attribute when deemed uncompilable.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>